+++ CHANGELOG.md
@@ -11,6 +11,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 - Added `ValidationContext` class to keep track of current validation path and errors up until a certain point.
+- Implemented non-default validation
+- Added dimension tests for `ZonString`
 
 ### Changed
 - Moved everything into a single file to combat circular reference issues